api/requirements.txt
PyJWT==2.1.0
flask==1.1.4
flask-restful==0.3.9
requests==2.25.1
python-dotenv==0.18.0